refactor(commands): extract target subcommands into target_cmds.rs
workspace_cmds.rs sat at 578 lines — under the 600-line ceiling but
close enough that the next substantive change risked forcing a second
round of splitting. The architecture audit flagged it alongside
skill_cmds.rs and persistence.rs as the three files pressuring that
limit.

Most of the bulk came from a single domain — target lifecycle — which
was structurally independent of workspace/binding/remote management:

  cmd_target             (dispatcher, 43 L)
  cmd_target_add         (95 L)
  cmd_target_remove      (69 L)

Move all three to src/commands/target_cmds.rs behind the same
`impl App` block. Rust's inherent-impl-across-modules pattern keeps the
API surface identical — callers in src/commands/mod.rs continue to
invoke `self.cmd_target(...)` without a module prefix.

Result:
  workspace_cmds.rs  578 → 366 L  (−37%, headroom for future edits)
  target_cmds.rs     new → 225 L  (focused single-domain module)

Scope decisions:
- Kept workspace_binding_add/remove in workspace_cmds.rs. They're
  workspace-level operations (pair an agent profile with a target via
  matcher) and share all the same helpers and mental model as
  cmd_status/cmd_doctor/cmd_remote.
- Did not touch cmd_status (74 L) or cmd_doctor (35 L) — they're
  workspace diagnostics, not target lifecycle.
- Re-registered the module in src/commands/mod.rs in alphabetical
  order (between sync_cmds and workspace_cmds).

Tested: make fmt-check, make lint (clippy -D warnings), make test
(75/75 pass, no flakes). The entire status/write/capture/reliability
integration suite — which constructs targets and asserts their
side-effects — passes unchanged, confirming the refactor is
behavior-preserving.
Scope-risk: narrow (3 files, net +14 lines from the new module's
imports and impl boilerplate).
Reversibility: clean — merge target_cmds.rs back into workspace_cmds
and drop the module declaration.
Confidence: high.
